Port timezone command to new timezone system
