+++ src/Command/index.ts
@@ -7,6 +7,7 @@ import {BuyCommand} from './Commands/BuyCommand';
 import {NotifyCommand} from './Commands/NotifyCommand';
 import {PricesCommand} from './Commands/PricesCommand';
 import {SellCommand} from './Commands/SellCommand';
+import {TimezoneCommand} from './Commands/TimezoneCommand';
 
 const PREFIX_REGEX = /:prefix\b/g;
 
@@ -16,6 +17,7 @@ export function init() {
 	COMMAND_MAP.add(new BuyCommand());
 	COMMAND_MAP.add(new SellCommand());
 	COMMAND_MAP.add(new PricesCommand());
+	COMMAND_MAP.add(new TimezoneCommand());
 	COMMAND_MAP.add(new NotifyCommand());
 	COMMAND_MAP.add(new HelpCommand());
